pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-wardrobe-tg-bot"
version = "0.1.0"
description = "AI Digital Wardrobe Telegram Bot and Mini App backend"
requires-python = ">=3.12"
dependencies = [
  "aioboto3>=13.2.0",
  "aiogram>=3.13.1",
  "alembic>=1.13.3",
  "asyncpg>=0.29.0",
  "celery[redis]>=5.4.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "orjson>=3.10.7",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.5.2",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.8",
  "sqlalchemy[asyncio]>=2.0.35",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.2",
  "pre-commit>=3.8.0",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.9",
  "types-PyJWT>=1.7.1",
]

[tool.setuptools.packages.find]
where = ["packages/core", "apps/api", "apps/bot", "apps/worker"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["apps", "packages", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "S", "RUF"]
ignore = ["S101", "S104", "S105", "S106", "RUF001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
disallow_untyped_decorators = false
warn_unreachable = true
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
mypy_path = ["packages/core", "apps/api", "apps/bot", "apps/worker"]
exclude = ["apps/miniapp", "alembic"]

[[tool.mypy.overrides]]
module = ["aioboto3.*", "aiogram.*", "celery.*", "fastapi.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
